+++ src/controllers/ProjectConfigController.php
@@ -8,8 +8,14 @@
 namespace craft\controllers;
 
 use Craft;
+use craft\helpers\FileHelper;
+use craft\helpers\ProjectConfig;
+use craft\helpers\StringHelper;
 use craft\web\Controller;
+use Symfony\Component\Yaml\Yaml;
+use yii\base\Exception;
 use yii\base\Response;
+use ZipArchive;
 
 /**
  * Manages the Project Config.
@@ -57,5 +63,34 @@ class ProjectConfigController extends Controller
         Craft::$app->getProjectConfig()->rebuild();
         $this->setSuccessFlash(Craft::t('app', 'Project config rebuilt successfully.'));
         return $this->redirectToPostedUrl();
+    }
+
+    /**
+     * Downloads the loaded project config as a zip file.
+     *
+     * @return Response
+     * @since 3.5.6
+     */
+    public function actionDownload(): Response
+    {
+        $config = Craft::$app->getProjectConfig()->get();
+        $splitConfig = ProjectConfig::splitConfigIntoComponents($config);
+        $zip = new ZipArchive();
+        $zipPath = Craft::$app->getPath()->getTempPath() . '/' . StringHelper::UUID() . '.zip';
+
+        if ($zip->open($zipPath, ZipArchive::CREATE) !== true) {
+            throw new Exception('Cannot create zip at ' . $zipPath);
+        }
+
+        foreach ($splitConfig as $path => $pathConfig) {
+            $content = Yaml::dump(ProjectConfig::cleanupConfig($pathConfig), 20, 2);
+            $zip->addFromString($path, $content);
+        }
+
+        $zip->close();
+        $this->response->sendContentAsFile(file_get_contents($zipPath), 'project.zip');
+        FileHelper::unlink($zipPath);
+
+        return $this->response;
     }
 }